PRMaps.makePolAngMap — Function

makePolAngMap(
    cam_ang::Stripeline.CameraAngles,
    signal::Healpix.PolarizedHealpixMap,
    setup::PRMaps.Setup
)

makePolAngMap(
    cam_ang::Stripeline.CameraAngles,
    signal::Healpix.PolarizedHealpixMap,
    setup::PRMaps.Setup,
    t_start::Dates.DateTime
)

makePolAngMap(
    cam_ang::Stripeline.CameraAngles,
    tel_ang::Stripeline.TelescopeAngles,
    signal::Healpix.PolarizedHealpixMap,
    setup::PRMaps.Setup
)

makePolAngMap(
    cam_ang::Stripeline.CameraAngles,
    tel_ang::Stripeline.TelescopeAngles,
    signal::Healpix.PolarizedHealpixMap,
    setup::PRMaps.Setup,
    t_start::Dates.DateTime
)

Return the polarization angle map of the sky observed by a telescope whose camera point towards a direction encoded by the CameraAngles struct.

The second flavour, with a TelescopeAngles as input, produce a map affected by an error. See makeErroredMap

Input:

cam_ang :: CameraAngles encoding the pointing direction of the detector;

tel_ang::Stripeline.TelescopeAngles encoding the non idealities of the telescope;

signal::Healpix.PolarizedHealpixMap the signal (Q,U,I) that the telescope are going to observe;

setup::PRMaps.Setup see Setup.

t_start::Dates.DateTime allows to set an initial date for the simulation.

Output:

p_map an HealpixMap containing observed angle of polarization.

PRMaps.makePolDegreeMap — Function

makePolDegreeMap(
    cam_ang::Stripeline.CameraAngles,
    signal::Healpix.PolarizedHealpixMap,
    setup::PRMaps.Setup
)

makePolDegreeMap(
    cam_ang::Stripeline.CameraAngles,
    signal::Healpix.PolarizedHealpixMap,
    setup::PRMaps.Setup,
    t_start::Dates.DateTime
)

makePolDegreeMap(
    cam_ang::Stripeline.CameraAngles,
    tel_ang::Stripeline.TelescopeAngles,
    signal::Healpix.PolarizedHealpixMap,
    setup::PRMaps.Setup
)

makePolDegreeMap(
    cam_ang::Stripeline.CameraAngles,
    tel_ang::Stripeline.TelescopeAngles,
    signal::Healpix.PolarizedHealpixMap,
    setup::PRMaps.Setup,
    t_start::Dates.DateTime
)

Return the polarization degree map of the sky observed by a telescope whose camera point towards a direction encoded by the CameraAngles struct.

The second flavour, with a TelescopeAngles as input, produce a map affected by an error. See makeErroredMap

Input:

cam_ang :: CameraAngles encoding the pointing direction of the detector;

tel_ang::Stripeline.TelescopeAngles encoding the non idealities of the telescope;

signal::Healpix.PolarizedHealpixMap the signal (Q,U,I) that the telescope are going to observe;

setup::PRMaps.Setup see Setup;

t_start::Dates.DateTime allows to set an initial date for the simulation.

Output:

p_map an HealpixMap containing observed degree of polarization.

PRMaps.polAngleMap! — Function

function polAngleMap!(
    ang_map::HealpixMap,
    q_map::HealpixMap,
    u_map::HealpixMap
)

This function calculate the polarization angle map and save it in ang_map. The angle of polarization is defined as:

0.5 * atan(U, Q)

Used internally by makePolAngMap.

Input:

ang_map::HealpixMap an empty HealpixMap

q_map::HealpixMap Q parameter HealpixMap

u_map::HealpixMap U parameter HealpixMap

Output:

nothing

PRMaps.polDegreeMap! — Function

function polDegreeMap!(
    p_map::HealpixMap,
    i_map::HealpixMap,
    q_map::HealpixMap,
    u_map::HealpixMap
)

This function calculate the polarization degree (not normalized) map and save it in ang_map. The degree of polarization is defined as:

sqrt(Q^2 + U^2)

Used internally by makePolAngMap.

Input:

p_map::HealpixMap an empty HealpixMap

i_map::HealpixMap I parameter HealpixMap

q_map::HealpixMap Q parameter HealpixMap

u_map::HealpixMap U parameter HealpixMap

Output:

nothing

PRMaps.differenceAngMaps — Function

differenceAngMaps(
    a :: HealpixMap,
    b :: HealpixMap
)

Calculate the difference between two maps containing the observed polarization angle.

Return an HealpixMap.

Angles MUST be expressed in RADIANS.