PRMaps.Setup — Type

Setup(
    sampling_freq_Hz :: Float64,
    total_time_s :: Float64    
)

Struct containing some useful data.

PRMaps.makeIdealMap — Function

makeIdealMap(
    cam_ang :: Sl.CameraAngles, 
    signal :: Healpix.HealpixMap,
    setup :: Setup
)

Generate a Healpix map using an ideal telescope model.

Return a tuple (map, hits)::(HealpixMap, HealpixMap):

map contains the obserbed values of signal;

hits contains for every pixels the count of how many times the pixel is seen.

makeIdealMap(
    cam_ang :: Sl.CameraAngles, 
    signal :: Healpix.HealpixMap,
    setup :: Setup
    t_start :: Dates.DateTime
)

Generate a Healpix map using an ideal telescope model.

Return a tuple (map, hits)::(HealpixMap, HealpixMap):

map contains the obserbed values of signal;

hits contains for every pixels the count of how many times the pixel is seen.

PRMaps.makeErroredMap — Function

makeErroredMap(
    cam_ang :: Sl.CameraAngles, 
    telescope_angles :: Sl.TelescopeAngles,
    signal :: Healpix.HealpixMap,
    setup :: Setup
)

Generate a Healpix map using a telescope model that takes into account the non idealities to generate the pointing direction. The observed values instead are calculated taking into account the ideal pointing directions. The result is a map affected by an error due to the non idealities of the system.

Return a tuple (map, hits)::(HealpixMap, HealpixMap):

map contains the obserbed values of signal;

hits contains for every pixels the count of how many times the pixel is seen.

makeErroredMap(
    cam_ang :: Sl.CameraAngles, 
    telescope_angles :: Sl.TelescopeAngles,
    signal :: Healpix.HealpixMap,
    setup :: Setup,
    t_start :: Dates.DateTime
)

Generate a Healpix map using a telescope model that takes into account the non idealities to generate the pointing direction. The observed values instead are calculated using the ideal pointing directions. The result is a map affected by an error due to the non idealities of the system.

Return a tuple (map, hits)::(HealpixMap, HealpixMap):

map contains the obserbed values of signal;

hits contains for every pixels the count of how many times the pixel is seen.

PRMaps.makeIdealMapIQU — Function

makeIdealMap(
    cam_ang :: Sl.CameraAngles, 
    signal :: Healpix.PolarizedHealpixMap,
    setup :: Setup
)

Generate a PolarizedHealpix map (I,Q,U) using an ideal telescope model.

Return a tuple (map, hits)::(PolarizedHealpixMap, PolarizedHealpixMap):

map contains the obserbed values of signal;

hits contains for every pixels the count of how many times the pixel is seen.

makeIdealMap(
    cam_ang :: Sl.CameraAngles, 
    signal :: Healpix.PolarizedHealpixMap,
    setup :: Setup,
    t_start :: Dates.DateTime
)

Generate a PolarizedHealpix map (I,Q,U) using an ideal telescope model.

Return a tuple (map, hits)::(PolarizedHealpixMap, PolarizedHealpixMap):

map contains the obserbed values of signal;

hits contains for every pixels the count of how many times the pixel is seen.

PRMaps.makeErroredMapIQU — Function

makeErroredMap(
    cam_ang :: Sl.CameraAngles, 
    telescope_angles :: Sl.TelescopeAngles,
    signal :: Healpix.PolarizedHealpixMap,
    setup :: Setup
)

Generate a PolarizedHealpix map (I,Q,U) using a telescope model that takes into account the non idealities to generate the pointing direction. The observed values instead are calculated taking into account the ideal pointing directions. The result is a map affected by an error due to the non idealities of the system.

Return a tuple (map, hits)::(PolarizedHealpixMap, PolarizedHealpixMap):

map contains the obserbed values of signal;

hits contains for every pixels the count of how many times the pixel is seen.

makeErroredMap(
    cam_ang :: Sl.CameraAngles, 
    telescope_angles :: Sl.TelescopeAngles,
    signal :: Healpix.PolarizedHealpixMap,
    setup :: Setup,
    t_start :: Dates.DateTime
)

Generate a PolarizedHealpix map (I,Q,U) using a telescope model that takes into account the non idealities to generate the pointing direction. The observed values instead are calculated taking into account the ideal pointing directions. The result is a map affected by an error due to the non idealities of the system.

Return a tuple (map, hits)::(PolarizedHealpixMap, PolarizedHealpixMap):

map contains the obserbed values of signal;

hits contains for every pixels the count of how many times the pixel is seen.